% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkm.R
\docType{package}
\name{mkm}
\alias{mkm}
\title{mkm: Support for reading structured spreadhseets.}
\description{
Provides the basis for a read() %>% parse() %>% validate()
%>% export() workflow for spreadsheets following the Mahinga Kai Monitoring (mkm) structure.
}
\section{Foo functions}{
 The foo functions ...
}

