% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mkm}
\alias{read_mkm}
\title{Read an MKM-structured spreadsheet}
\usage{
read_mkm(path, section_text, sheet = 1)
}
\arguments{
\item{path}{A string. The location of an \code{xlsx}, \code{xls} or \code{csv}
files to read.}

\item{section_text}{A (optionally named) character vector or scalar. See Details}

\item{sheet}{The sheet to read. Either a string (the name of a sheet), or an
integer (the position of the sheet). Ignored for \code{csv} files.}
}
\value{
A named list with \code{length(sections)} items representing the
 sections extracted from the file. Each item is a list with two sections:

 \itemize{

 \item{\code{dat}: Data extracted from the section}

 \item{\code{format}: Formatting of the data in the section. Always
 \code{NULL} for \code{.xls} and \code{.csv} files} because formatting cannot
 be read from these file types.

 }
}
\description{
Read an MKM-structured spreadsheet
}
\details{
This function reads either an \code{xlsx}, \code{xls} or \code{csv}
 file, searches the first (non-empty) column for the text contained in
 \code{section_text} and returns an object with various 'sections' that can then
 be further parsed into sensible R data structures (e.g. via
 \code{\link{parse_mkm}}).

 Sections are denoted by one or more plain text strings (\code{search_text}). If
 these are supplied as a named character vector, these names are used for the
 sections in the returned list. Where one or more names are missing or blank
 the search text is used and all names are made syntactically valid and unique (via
 \code{\link{make_keys}}).
}
\examples{

}
