% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to...}
\alias{to...}
\alias{to_lst}
\alias{to_tlst}
\alias{to_df}
\alias{to_tdf}
\alias{to_mat}
\alias{to_tmat}
\alias{to_meta}
\title{Coerce MKM objects to base R types}
\usage{
to_lst(x)

to_tlst(x)

to_df(x)

to_tdf(x)

to_mat(x)

to_tmat(x)

to_meta(x)
}
\arguments{
\item{x}{An MKM object, similar to that produced by \code{\link{read_mkm}}}
}
\value{

}
\description{
\code{to_*} functions take raw input (which is a list with 'dat' and 'format' components) and produce a base R type (list, data.frame or matrix) or something slightly special.
}
\details{
Using custom functions, \code{MKM} can produce any data structure. However, support for three fundamental types are provided.

\itemize{

\item \code{to_lst } : Produces a list by reading names from column headers and disregarding any row names.
\item \code{to_tlst} : Produces a list using names from row headers (and disregarding column headers)
\item \code{to_df  } : Produces a data.frame by reading names from column headers and creating an \code{.id} column to hold information in row headers (if any).
\item \code{to_tdf } : Produces a data.frame using names from row headers and creating an \code{.id} column to hold information in column headers (if any).
\item \code{to_mat } : Produces a matrix, with \code{dimnames} set from row and column headers in the direction they appear.
\item \code{to_tmat} : Same as \code{to_mat} but transposed.

}
}
