% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{to_lst}
\alias{to_lst}
\alias{to_tlst}
\alias{to_meta}
\alias{to_df}
\alias{to_tdf}
\alias{to_mat}
\alias{to_tmat}
\title{Coerce MKM objects to base R types}
\usage{
to_lst(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
to_* functions take raw input (which is a list with 'text' and 'format' components) and produce a base R type (list, dataframe or matrix)
}
