% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mkm}
\alias{parse_mkm}
\title{Parse}
\usage{
parse_mkm(x, definition = NULL, .default = "to_df")
}
\arguments{
\item{x}{An MKM object, similar to that produced by \code{\link{read_mkm}}}

\item{definition}{A definition object}

\item{.default}{The defult parsing function to use, if none is specified}
}
\value{

}
\description{
Parse
}
